+++ tests/event/fn/project-event.test.ts
@@ -1,8 +1,7 @@
 import { describe, expect, it } from 'bun:test'
+import type { ExtendedDomainEvent } from '../../../src'
+import { ReadDatabaseInMemory, id } from '../../../src'
 import { createProjectEventFnFactory } from '../../../src/event/fn/project-event'
-import type { ExtendedDomainEvent } from '../../../src/types'
-import { ReadDatabaseInMemory } from '../../../src/utils'
-import { id } from '../../../src/types'
 import type { CounterEvent } from '../../data/command/counter'
 import { counterReactor } from '../../data/command/counter'
 import type { CounterView } from '../../data/query/view'
