+++ tests/utils/command-dispatcher-mock.test.ts
@@ -1,6 +1,5 @@
 import { describe, expect, it } from 'bun:test'
-import { id } from '../../src'
-import { CommandDispatcherMock } from '../../src/utils/command-dispatcher-mock'
+import { CommandDispatcherMock, id } from '../../src'
 
 describe('command dispatcher mock', () => {
   describe('dispatch', () => {
@@ -19,7 +18,7 @@ describe('command dispatcher mock', () => {
       const command = {
         type: 'counter',
         operation: 'increment',
-        id: { type: 'counter', id: '1' }
+        id: id('counter', '1')
       }
 
       await dispatcher.dispatch(command)
@@ -39,11 +38,11 @@ describe('command dispatcher mock', () => {
       const dispatcher = new CommandDispatcherMock()
       const command1 = {
         operation: 'create',
-        id: { type: 'counter', id: '1' }
+        id: id('counter', '1')
       }
       const command2 = {
         operation: 'increment',
-        id: { type: 'counter', id: '1' }
+        id: id('counter', '1')
       }
       await dispatcher.dispatch(command1)
       await dispatcher.dispatch(command2)
@@ -56,7 +55,7 @@ describe('command dispatcher mock', () => {
       const dispatcher = new CommandDispatcherMock()
       const command = {
         operation: 'create',
-        id: { type: 'counter', id: '1' }
+        id: id('counter', '1')
       }
       await dispatcher.dispatch(command)
       dispatcher.reset()
@@ -68,7 +67,7 @@ describe('command dispatcher mock', () => {
       const command = {
         type: 'counter',
         operation: 'increment',
-        id: { type: 'counter', id: '1' }
+        id: id('counter', '1')
       }
 
       await dispatcher.dispatch(command)
